+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### 3.4.8 (pending)
 
+* ldap weaks
+  * dashboard sso is no configured with `sso` in `dashboard.properties`
+  * ldap group (for ldap lookup and dashboard sso) is now configurable via ldap.sso 
+    (ldap.sso will be renamed to ldap.group in stool 3.5; it's not renamed now to avoid migration problems)
 * dumped `system-start` and `system-stop`, use `start -all`  and `stop -all` instead
 * fixed url defaults, they never worked (thanks to Max)
 * adjust autoconf url for cp: replaced /xml/config by (/internal-login)